separate distribution endpoints and router for clarity
