+++ packages/api/endpoints/distribution.js
@@ -1,16 +1,13 @@
 'use strict';
 
-const router = require('express-promise-router')();
 const urljoin = require('url-join');
 const {
   s3,
   getFileBucketAndKey
 } = require('@cumulus/common/aws');
-const { randomId } = require('@cumulus/common/test-utils');
 const { UnparsableFileLocationError } = require('@cumulus/common/errors');
 const { URL } = require('url');
 const EarthdataLogin = require('../lib/EarthdataLogin');
-const { RecordDoesNotExist } = require('../lib/errors');
 const { isLocalApi } = require('../lib/testUtils');
 const { AccessToken } = require('../models');
 const s3credentials = require('./s3credentials');
@@ -40,16 +37,6 @@ function getSignedS3Url(s3Client, Bucket, Key, username) {
   parsedSignedUrl.searchParams.set('x-EarthdataLoginUsername', username);
 
   return parsedSignedUrl.toString();
-}
-
-/**
- * Checks if the token is expired
- *
- * @param {Object} accessTokenRecord - the access token record
- * @returns {boolean} true indicates the token is expired
- */
-function isAccessTokenExpired(accessTokenRecord) {
-  return accessTokenRecord.expirationTime < Date.now();
 }
 
 /**
@@ -111,51 +98,6 @@ async function handleRedirectRequest(req, res) {
 }
 
 /**
- * Ensure request is authorized through EarthdataLogin or redirect to become so.
- *
- * @param {Object} req - express request object
- * @param {Object} res - express response object
- * @param {Function} next - express middleware callback function
- * @returns {Promise<Object>} - promise of an express response object
- */
-async function ensureAuthorizedOrRedirect(req, res, next) {
-  // Skip authentication for debugging purposes.
-  if (process.env.FAKE_AUTH) {
-    req.authorizedMetadata = { userName: randomId('username') };
-    return next();
-  }
-
-  const {
-    accessTokenModel,
-    authClient
-  } = getConfigurations();
-
-  const redirectURLForAuthorizationCode = authClient.getAuthorizationUrl(req.path);
-  const accessToken = req.cookies.accessToken;
-
-  if (!accessToken) return res.redirect(307, redirectURLForAuthorizationCode);
-
-  let accessTokenRecord;
-  try {
-    accessTokenRecord = await accessTokenModel.get({ accessToken });
-  }
-  catch (err) {
-    if (err instanceof RecordDoesNotExist) {
-      return res.redirect(307, redirectURLForAuthorizationCode);
-    }
-
-    throw err;
-  }
-
-  if (isAccessTokenExpired(accessTokenRecord)) {
-    return res.redirect(307, redirectURLForAuthorizationCode);
-  }
-
-  req.authorizedMetadata = { userName: accessTokenRecord.username };
-  return next();
-}
-
-/**
  * Responds to a request for temporary s3 credentials.
  *
  * @param {any} req - express request object
@@ -202,8 +144,8 @@ async function handleFileRequest(req, res) {
     .send('Redirecting');
 }
 
-router.get('/redirect', handleRedirectRequest);
-router.get('/s3credentials', ensureAuthorizedOrRedirect, handleCredentialRequest);
-router.get('/*', ensureAuthorizedOrRedirect, handleFileRequest);
-
-module.exports = router;
+module.exports = {
+  handleRedirectRequest,
+  handleCredentialRequest,
+  handleFileRequest
+};
